requirements.txt
tensorflow
opencv-python
numpy
mysql-connector-python
